requirements.txt
python-dotenv==0.15.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
PyMySQL
coverage
flask_sqlalchemy
pyyaml
nose2
flask-cors
flake8
cryptography
Flask-Migrate
mwoauth
flask_login
wikibase-api
wikidata